+++ scripts/src/blocks/RandomBlock.ts
@@ -9,7 +9,10 @@ class RandomBlock extends Block {
         this.max = max;
         this.variableName = settings.defaultVariablePrefix + (globalVariables.size+1);
 
-        globalVariables.set(this.variableName, null);
+        if(globalVariables.has(this.variableName) == false)
+        {
+            globalVariables.set(this.variableName, null);
+        }
         this.init();
     }