+++ scripts/src/blocks/StartBlock.ts
@@ -1,13 +1,15 @@
 class StartBlock extends Block {
-    runSpeed: number = 500;
+    runSpeed: number;
     isDeletable: boolean = false;
-    constructor(x: number = 0, y: number = 0) {
+    constructor(x: number = 0, y: number = 0, runSpeed: number = 500) {
         super(x, y);
+        this.runSpeed = runSpeed;
         this.init();
     }
 
     createBlock(): void {
         workspace.innerHTML += `<div class="block start" id="${this.id}">Start</div>`;
+        runSpeed = this.runSpeed;
     }
 
     properties(): void {
@@ -15,7 +17,7 @@ class StartBlock extends Block {
             if(e.button == 1) return;
             
             propertiesWindow.innerHTML = `
-                <p>Run Speed (ms): <input type="number" value="${runSpeed}" class="property${this.id}"></p>
+                <p>Run Speed (ms): <input type="number" value="${this.runSpeed}" class="property${this.id}"></p>
             `;
 
             const property: any = propertiesWindow.querySelectorAll(".property"+this.id);